refactor(kosong): split Anthropic and OpenAI Responses providers
